+++ frontend/src/views/CatalogView.tsx
@@ -95,6 +95,8 @@ const CatalogList = ({ catalog, title }: CatalogListProps) => {
             ]}
             rows={catalog}
             getRowId={(row) => row.assetId}
+            getRowHeight={() => "auto"}
+            disableRowSelectionOnClick
         ></Table>
     );
 };