+++ CHANGELOG.md
@@ -17,7 +17,7 @@ The **need for configuration updates** is **marked bold**.
 
 ### Changed
 
-- /
+- Updated table row height and selection behavior ([#1135](https://github.com/eclipse-tractusx/puris/pull/1135))
 
 ### Fixes